requirements.txt
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.3.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tensorflow>=2.8.0
umap-learn>=0.5.0
jupyter>=1.0.0
ipython>=7.0.0